pyproject.toml
[tool.poetry]
name = "smartkitchen-api"
version = "0.1.0"
description = ""
authors = ["Lucas Martins <lucasmartins1709@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
fastapi = "^0.111.0"
motor = "^3.4.0"
beanie = "^1.26.0"
pydantic-settings = "^2.3.4"
pwdlib = { extras = ["argon2"], version = "^0.2.0" }
pyjwt = "^2.8.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
setuptools = "^72.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.4.10"
pytest-asyncio = "^0.23.7"
commitizen = "^3.27.0"
asgi-lifespan = "^2.1.0"
mongomock = "^4.1.2"
mongomock-motor = "^0.0.31"
faker = "^26.2.0"
freezegun = "^1.5.1"

[tool.ruff]
# line-length = 79
# extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_mode = "strict"


[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev smartkitchen_api/main.py'
pre_test = 'task lint'
post_test = 'coverage html'
test = 'pytest -s --cov=smartkitchen_api -vv'
# O -x no pytest faz com que o teste pare no primeiro error.
# O -s mostra as saídas no console, bom para usar print().
# O --pdb entra em modo DEBBUG

[[tool.mypy.overrides]]
module = "mongomock_motor"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
